refactor(ingestPutController.ts): remove unused import and parameter from sendProviderChangeDetailsEmailToAdmins function refactor(govuk-notify.ts): move import statements to the top and add List type import
